refactor: nicely segregate the config
This commit does a complete overhaul of the config structure. It aims to
introduce simplicity with the ease of maintainance by modularizing stuff
very strictly.

For example, to set up keymaps on plugin load; instead of declaring them
in the load function, declare them in the keymaps and call it from the
load.